-fix: Various
Minimaps now show "correctly" for both 1.2 and 3.5 clients
Added some cleanup code to save some memory when closing the map
Closing the map now just hides it. As a downside you might need to restart the viewer when you change DB or game_pak before things are updated.
Sphere markers now show as circles on the map. (still at the wrong location for quests)
Added a button to show housing
Added a error message when a gale_pak fails to open (this is usually because of out of memory error)
Some map view settings are now saved