-add: customizable popup menu for map
You can now use customizable popups from the map viewer when you right-click it.
If you click on a sub-menu item, it will copy the related info to clipboard. The menus can be customized for every type of Point of Interest added.
Check the example menus in data/*.menu and the source at MapViewForm.AddMenuItems() for more info.