-add: treeview node double-click
You can now double-click on properties in the Treeview for doodads and quests workflow to copy the entry to clipboard.
When it is a entry that has auto-completed values on it, like for items, skills, etc. it will also open the search-tab for the selected entry and search for it.